Moved class output to separate class
